# 切割后的立方体reshape为net输入状态 # 微调ballproposal参数 # 整合negcubes为neg0/neg1/..
# cnn3d
1. net版本完成搭建
2. 初试效果不好
3. 将尝试不同大小node分开训练网络
4. 尝试cube更改截取方式（按照等效半径截取）

+++ tutorial_code/imgpreview.py
@@ -85,7 +85,8 @@ elif mode == '2':
     while True:
         img_file = input(working_path+'{文件}：')
         img_file = working_path+img_file
-        imgs = np.load(img_file).astype(np.float32)
+        # imgs = np.load(img_file).astype(np.float32)
+        imgs = np.load(img_file)
         plt.subplot(131)
         plt.imshow(imgs[0])
         plt.colorbar()
